+++ packages/vite-plugin/src/node/manifest.ts
@@ -166,12 +166,12 @@ export interface ManifestV3 {
   incognito?: string | undefined
   input_components?:
     | {
-        name?: string | undefined
-        type?: string | undefined
+        name: string
         id?: string | undefined
-        description?: string | undefined
-        language?: string | undefined
-        layouts?: string[] | undefined
+        language?: string | string[] | undefined
+        layouts?: string | string[] | undefined
+        input_view?: string | undefined
+        options_page?: string | undefined
       }[]
     | undefined
   key?: string | undefined
@@ -236,6 +236,11 @@ export interface ManifestV3 {
     | {
         pages: string[]
         content_security_policy?: string | undefined
+      }
+    | undefined
+  side_panel?:
+    | {
+      default_path?: string | undefined
       }
     | undefined
   short_name?: string | undefined