+++ packages/vite-plugin/tests/e2e/mv3-vite-react-content-script-seq-hmr/src1/A.jsx
@@ -1,0 +1,8 @@
+import { C } from './C'
+
+export const A = () => (
+  <div>
+    <p className='A'>a-0</p>
+    <C />
+  </div>
+)