+++ packages/vite-plugin/tests/e2e/mv3-vite-react-content-script-seq-hmr/src1/App.jsx
@@ -1,0 +1,13 @@
+import { A } from './A'
+import { B } from './B'
+
+function App() {
+  return (
+    <div id='app'>
+      <A />
+      <B />
+    </div>
+  )
+}
+
+export default App