+++ packages/vite-plugin/tests/e2e/mv3-vite-svelte-page/vite-build.test.ts
@@ -2,14 +2,18 @@ import { getCustomId, getPage } from '../helpers'
 import { build } from '../runners'
 import { test, expect } from 'vitest'
 
-test('crx runs from build output', async (ctx) => {
-  const { browser } = await build(__dirname)
-  const page = await getPage(browser, 'chrome-extension')
+test(
+  'crx runs from build output',
+  async (ctx) => {
+    const { browser } = await build(__dirname)
+    const page = await getPage(browser, 'chrome-extension')
 
-  const app = page.locator('#app')
-  await app.locator('img').waitFor()
+    const app = page.locator('#app')
+    await app.locator('img').waitFor()
 
-  expect(await app.screenshot()).toMatchImageSnapshot({
-    customSnapshotIdentifier: getCustomId(ctx),
-  })
-})
+    expect(await app.screenshot()).toMatchImageSnapshot({
+      customSnapshotIdentifier: getCustomId(ctx),
+    })
+  },
+  { retry: 2 },
+)