+++ packages/vite-plugin/tests/e2e/mv3-vite-vanilla-content-script-seq-hmr/src2/b.ts
@@ -1,0 +1,9 @@
+import { c1 } from './c1'
+import { c2 } from './c2'
+
+export const b = 'second'
+
+export const c = {
+  c1,
+  c2,
+}