+++ packages/vite-plugin/tests/e2e/mv3-web-accessible-resources/vite-serve.test.ts
@@ -16,5 +16,5 @@ test(
     await image.waitFor({ state: 'attached' })
     await script.waitFor({ state: 'attached' })
   },
-  { retry: process.env.CI ? 5 : 0 },
+  { retry: process.env.CI ? 5 : 2 },
 )